Move createDateFromTimestamp function to a general date/time utils file for better organization
