+++ advisor-ui/src/utils/dateTimeUtils.js
@@ -1,0 +1,9 @@
+const createDateFromTimeStamp = (timestamp) => {
+  const date = new Date(timestamp);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  return `${year}-${month}-${day}`;
+};
+
+export { createDateFromTimeStamp };
